Rework mage
