+++ magefiles/utils/helm.go
@@ -1,5 +1,3 @@
-//go:build mage
-
 /*
 Copyright 2023 The Kubernetes Authors.
 
@@ -16,178 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
-
-import (
-	"bytes"
-	"fmt"
-	"os"
-	"strings"
-
-	semver "github.com/blang/semver/v4"
-	"github.com/helm/helm/pkg/chartutil"
-	"github.com/magefile/mage/mg"
-	"github.com/magefile/mage/sh"
-	yamlpath "github.com/vmware-labs/yaml-jsonpath/pkg/yamlpath"
-	"gopkg.in/yaml.v3"
-)
-
-const (
-	HelmChartPath   = "charts/ingress-nginx/Chart.yaml"
-	HelmChartValues = "charts/ingress-nginx/values.yaml"
-)
-
-type Helm mg.Namespace
-
-// UpdateAppVersion Updates the Helm App Version of Ingress Nginx Controller
-func (Helm) UpdateAppVersion() {
-	updateAppVersion()
-}
-
-func updateAppVersion() {
-}
-
-// UpdateVersion Update Helm Version of the Chart
-func (Helm) UpdateVersion(version string) {
-	updateVersion(version)
-}
-
-func currentChartVersion() string {
-	chart, err := chartutil.LoadChartfile(HelmChartPath)
-	CheckIfError(err, "HELM Could not Load Chart")
-	return chart.Version
-}
-
-func currentChartAppVersion() string {
-	chart, err := chartutil.LoadChartfile(HelmChartPath)
-	CheckIfError(err, "HELM Could not Load Chart")
-	return chart.AppVersion
-}
-
-func updateVersion(version string) {
-	Info("HELM Reading File %v", HelmChartPath)
-
-	chart, err := chartutil.LoadChartfile(HelmChartPath)
-	CheckIfError(err, "HELM Could not Load Chart")
-
-	// Get the current tag
-	// appVersionV, err := getIngressNGINXVersion()
-	// CheckIfError(err, "HELM Issue Retrieving the Current Ingress Nginx Version")
-
-	// remove the v from TAG
-	appVersion := version
-
-	Info("HELM Ingress-Nginx App Version: %s Chart AppVersion: %s", appVersion, chart.AppVersion)
-	if appVersion == chart.AppVersion {
-		Warning("HELM Ingress NGINX Version didnt change Ingress-Nginx App Version: %s Chart AppVersion: %s", appVersion, chart.AppVersion)
-		return
-	}
-
-	// Update the helm chart
-	chart.AppVersion = appVersion
-	cTag, err := semver.Make(chart.Version)
-	CheckIfError(err, "HELM Creating Chart Version: %v", err)
-
-	if err = cTag.IncrementPatch(); err != nil {
-		ErrorF("HELM Incrementing Chart Version: %v", err)
-		os.Exit(1)
-	}
-	chart.Version = cTag.String()
-	Debug("HELM Updated Chart Version: %v", chart.Version)
-
-	err = chartutil.SaveChartfile(HelmChartPath, chart)
-	CheckIfError(err, "HELM Saving new Chart")
-}
-
-func updateChartReleaseNotes(releasesNotes []string) {
-	Info("HELM Updating the Chart Release notes")
-	chart, err := chartutil.LoadChartfile(HelmChartPath)
-	CheckIfError(err, "HELM Could not Load Chart to update release notes %s", HelmChartPath)
-	var releaseNoteString string
-	for i := range releasesNotes {
-		releaseNoteString = fmt.Sprintf("%s - \"%s\"\n", releaseNoteString, releasesNotes[i])
-	}
-	Info("HELM Release note string %s", releaseNoteString)
-	chart.Annotations["artifacthub.io/changes"] = releaseNoteString
-	err = chartutil.SaveChartfile(HelmChartPath, chart)
-	CheckIfError(err, "HELM Saving updated release notes for Chart")
-}
-
-func UpdateChartChangelog() {
-}
-
-// UpdateChartValue Updates the Helm ChartValue
-func (Helm) UpdateChartValue(key, value string) {
-	updateChartValue(key, value)
-}
-
-func updateChartValue(key, value string) {
-	Info("HELM Updating Chart %s %s:%s", HelmChartValues, key, value)
-
-	// read current values.yaml
-	data, err := os.ReadFile(HelmChartValues)
-	CheckIfError(err, "HELM Could not Load Helm Chart Values files %s", HelmChartValues)
-
-	// var valuesStruct IngressChartValue
-	var n yaml.Node
-	CheckIfError(yaml.Unmarshal(data, &n), "HELM Could not Unmarshal %s", HelmChartValues)
-
-	// update value
-	// keyParse := parsePath(key)
-	p, err := yamlpath.NewPath(key)
-	CheckIfError(err, "HELM cannot create path")
-
-	q, err := p.Find(&n)
-	CheckIfError(err, "HELM unexpected error finding path")
-
-	for _, i := range q {
-		Info("HELM Found %s at %s", i.Value, key)
-		i.Value = value
-		Info("HELM Updated %s at %s", i.Value, key)
-	}
-
-	//// write to file
-	var b bytes.Buffer
-	yamlEncoder := yaml.NewEncoder(&b)
-	yamlEncoder.SetIndent(2)
-	err = yamlEncoder.Encode(&n)
-	CheckIfError(err, "HELM Could not Marshal new Values file")
-	err = os.WriteFile(HelmChartValues, b.Bytes(), 0o644)
-	CheckIfError(err, "HELM Could not write new Values file to %s", HelmChartValues)
-
-	Info("HELM Ingress Nginx Helm Chart update %s %s", key, value)
-}
-
-func (Helm) Helmdocs() error {
-	return runHelmDocs()
-}
-
-func runHelmDocs() error {
-	err := installHelmDocs()
-	if err != nil {
-		return err
-	}
-	err = sh.RunV("helm-docs", "--chart-search-root=${PWD}/charts")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func installHelmDocs() error {
-	Info("HELM Install HelmDocs")
-	g0 := sh.RunCmd("go")
-
-	err := g0("install", "github.com/norwoodj/helm-docs/cmd/helm-docs@v1.11.0")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-func parsePath(key string) []string { return strings.Split(key, ".") }
-
-func updateHelmDocs() {
-}
+package utils
 
 type IngressChartValue struct {
 	CommonLabels struct{} `yaml:"commonLabels"`
