+++ magefiles/utils/templates/e2edocs.tpl
@@ -1,0 +1,10 @@
+<!---
+This file is autogenerated!
+Do not try to edit it manually.
+-->
+
+# e2e test suite for [Ingress NGINX Controller]({{.URL}})
+
+{{ range $test := .Tests }}
+{{ $test }}
+{{- end }}
